Implementing DI on UserController-AddUser
