+++ views/addUser.twig
@@ -1,6 +1,6 @@
 {% extends "layout.twig"%} {% block content %}
 
-<h5 class="card-title">Crear Usuario</h5>
+<h5 class="card-title">Create User</h5>
 
 <div class="alert alert-primary" role="alert">
   {{ responseMessage }}
